code restructure, made a buffers file
