+++ src/main.rs
@@ -1,7 +1,9 @@
 mod Scene;
+mod buffers;
 mod camera;
 mod renderer;
 
+use buffers::{Params, SceneMaterial, SceneSphere};
 use camera::Camera;
 use Scene::{Material, RenderScene, Sphere};
 
@@ -10,9 +12,9 @@ use egui::{pos2, Color32, DragValue, Frame, FullOutput};
 use renderer::Renderer;
 
 use wgpu::{
-    include_wgsl, util::DeviceExt, Adapter, Backends, BindGroup, Buffer, Device, Dx12Compiler,
-    Gles3MinorVersion, Instance, InstanceDescriptor, InstanceFlags, PipelineLayout, Queue, Surface,
-    TextureDescriptor, TextureDimension, TextureFormat, TextureUsages,
+    include_wgsl, Adapter, Backends, BindGroup, Device, Dx12Compiler, Gles3MinorVersion, Instance,
+    InstanceDescriptor, InstanceFlags, PipelineLayout, Queue, Surface, TextureDescriptor,
+    TextureDimension, TextureFormat, TextureUsages,
 };
 
 use winit::{
@@ -28,53 +30,6 @@ use egui_winit_platform::{Platform, PlatformDescriptor};
 use glam::vec3a;
 
 use std::time::Instant;
-
-use futures::executor::block_on;
-
-fn vec3_pad(x: f32, y: f32, z: f32) -> [f32; 4] {
-    [x, y, z, 0.0]
-}
-
-#[repr(C)]
-#[derive(Debug, Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
-struct Params {
-    width: u32,              // float, aligned to 4 bytes
-    accumulation_index: u32, // u32, aligned to 4 bytes
-    _padding: [u8; 8],       // padding to ensure 16-byte alignment
-}
-
-#[repr(C)]
-#[derive(Debug, Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
-struct RayCamera {
-    pub origin: [f32; 4],    // vec4, aligned to 16 bytes
-    pub direction: [f32; 4], // vec4, aligned to 16 bytes
-}
-
-#[repr(C)]
-#[derive(Debug, Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
-struct Ray {
-    pub direction: [f32; 4], // vec4, aligned to 16 bytes
-}
-
-#[repr(C)]
-#[derive(Debug, Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
-struct SceneSphere {
-    pub position: [f32; 3],  // vec3, aligned to 12 bytes
-    pub radius: f32,         // f32, aligned to 4 bytes
-    pub material_index: u32, // u32, aligned to 4 bytes
-    _padding: [u8; 12],      // padding to ensure 16-byte alignment
-}
-
-#[repr(C)]
-#[derive(Debug, Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
-struct SceneMaterial {
-    pub albedo: [f32; 3],         // vec3, aligned to 12 bytes
-    pub roughness: f32,           // f32, aligned to 4 bytes
-    pub emission_color: [f32; 3], // vec3, aligned to 12 bytes
-    pub metallic: f32,            // f32, aligned to 4 bytes
-    pub emission_power: f32,      // f32, aligned to 4 bytes
-    _padding: [u8; 12],           // padding to ensure 16-byte alignment
-}
 
 pub fn main() {
     let event_loop = EventLoop::new().unwrap();
@@ -260,8 +215,6 @@ async fn run(event_loop: EventLoop<()>, window: Window) {
 
     let scene = define_scene();
 
-    let mut scene_renderer = Renderer::new(camera.clone(), scene);
-
     let mut last_mouse_pos: egui::Pos2 = pos2(0., 0.);
 
     let instance = generate_instance();
@@ -271,50 +224,25 @@ async fn run(event_loop: EventLoop<()>, window: Window) {
     // Create the logical device and command queue
     let (device, queue) = generate_device_and_queue(&adapter).await;
 
-    // ################################ GPU DATA PIPELINE #########################################
-
-    let mut camera_rays: Vec<Ray> = generate_new_camera_rays(&camera);
-
     let (material_array, sphere_array) = define_render_scene();
     // Create uniform buffer
-    let mut params = Params {
+    let params = Params {
         width: size.width,
         accumulation_index: 1,
         _padding: [0; 8],
     };
 
-    let (
-        mut output_buffer_size,
-        mut accumulation_buffer_size,
-        mut input_buffer,
-        mut output_buffer,
-        mut params_buffer,
-        mut staging_buffer,
-        mut camera_buffer,
-        mut material_buffer,
-        mut sphere_buffer,
-        mut accumulation_buffer,
-    ) = create_buffers(
+    let (mut scene_renderer, compute_bindgroup_layout, compute_bind_group) = Renderer::new(
+        camera,
+        scene,
         &device,
         &size,
-        &camera_rays,
         &material_array,
         &sphere_array,
         &[params],
     );
 
-    let mut bytes_per_row = calculate_bytes_per_row(&size);
-
-    let (mut compute_bindgroup_layout, mut compute_bind_group) = create_compute_bindgroup(
-        &device,
-        &input_buffer,
-        &output_buffer,
-        &params_buffer,
-        &camera_buffer,
-        &material_buffer,
-        &sphere_buffer,
-        &accumulation_buffer,
-    );
+    // ################################ GPU DATA PIPELINE #########################################
 
     // #####################################################################################
 
@@ -439,41 +367,7 @@ async fn run(event_loop: EventLoop<()>, window: Window) {
                                 create_device_bindgroup(&device, &texture, &sampler);
 
                             // ###################### COMPUTE RESIZING ##################################
-                            camera_rays = generate_new_camera_rays(&camera);
 
-                            (
-                                output_buffer_size,
-                                accumulation_buffer_size,
-                                input_buffer,
-                                output_buffer,
-                                params_buffer,
-                                staging_buffer,
-                                camera_buffer,
-                                material_buffer,
-                                sphere_buffer,
-                                accumulation_buffer,
-                            ) = create_buffers(
-                                &device,
-                                &size,
-                                &camera_rays,
-                                &material_array,
-                                &sphere_array,
-                                &[params],
-                            );
-
-                            bytes_per_row = calculate_bytes_per_row(&size);
-
-                            (compute_bindgroup_layout, compute_bind_group) =
-                                create_compute_bindgroup(
-                                    &device,
-                                    &input_buffer,
-                                    &output_buffer,
-                                    &params_buffer,
-                                    &camera_buffer,
-                                    &material_buffer,
-                                    &sphere_buffer,
-                                    &accumulation_buffer,
-                                );
                             // ####################################################
 
                             surface.configure(&device, &surface_config);
@@ -559,10 +453,6 @@ async fn run(event_loop: EventLoop<()>, window: Window) {
 
                             // ###################################### update data step ########################################
 
-                            params.accumulation_index += 1;
-
-                            queue.write_buffer(&params_buffer, 0, bytemuck::cast_slice(&[params]));
-
                             /*let mut buffer_encoder =
                                 device.create_command_encoder(&wgpu::CommandEncoderDescriptor {
                                     label: Some("Buffer Encoder"),
@@ -577,118 +467,14 @@ async fn run(event_loop: EventLoop<()>, window: Window) {
                             queue.submit(Some(buffer_encoder.finish()));*/
 
                             // #############################################################################################
-                            // ###################################### compute step ########################################
 
-                            let mut compute_encoder =
-                                device.create_command_encoder(&wgpu::CommandEncoderDescriptor {
-                                    label: Some("Compute Encoder"),
-                                });
-
-                            {
-                                let mut compute_pass = compute_encoder.begin_compute_pass(
-                                    &wgpu::ComputePassDescriptor {
-                                        label: Some("Compute Pass"),
-                                        timestamp_writes: None,
-                                    },
-                                );
-                                compute_pass.set_pipeline(&compute_pipeline);
-                                compute_pass.set_bind_group(0, &compute_bind_group, &[]);
-                                compute_pass.dispatch_workgroups(
-                                    (size.width + 7) / 8,
-                                    (size.height + 7) / 8,
-                                    1,
-                                );
-                            }
-
-                            let test = false;
-
-                            if test {
-                                // Copy the data from the output buffer to the staging buffer
-                                compute_encoder.copy_buffer_to_buffer(
-                                    &output_buffer,
-                                    0,
-                                    &staging_buffer,
-                                    0,
-                                    output_buffer_size,
-                                )
-                            }
-
-                            queue.submit(Some(compute_encoder.finish()));
-
-                            //println!(
-                            //    "time 2: {}",
-                            //    start_time.elapsed().as_micros() as f32 / 1000.
-                            //);
-
-                            // #############################################################################################
-
-                            // ###################################### new texture from buffers ########################################
-
-                            if test {
-                                let buffer_slice = staging_buffer.slice(..);
-                                let (sender, receiver) = futures::channel::oneshot::channel();
-
-                                buffer_slice.map_async(wgpu::MapMode::Read, move |result| {
-                                    sender.send(result).unwrap();
-                                });
-
-                                device.poll(wgpu::Maintain::Wait);
-
-                                let pixel_colors: Vec<u8>;
-                                // Wait for the mapping to complete
-                                if let Ok(Ok(())) = block_on(receiver) {
-                                    // Read the buffer data
-                                    let data = buffer_slice.get_mapped_range();
-
-                                    // Cast the data to f32 and convert it to u8
-                                    pixel_colors = bytemuck::cast_slice::<u8, u8>(&data).to_vec();
-
-                                    //println!(
-                                    //    "wanted: {}, actual: {}",
-                                    //    size.height * size.width * 4,
-                                    //    pixel_colors.len()
-                                    //);
-
-                                    println!("pixel colors: {:?}", pixel_colors[0]);
-                                } else {
-                                    panic!("Failed to map buffer");
-                                }
-                                // Unmap the buffer
-                                staging_buffer.unmap();
-                            }
-
-                            // #############################################################################################
-
-                            // ###################################### copying buffers to texture ########################################
-                            let mut copy_encoder =
-                                device.create_command_encoder(&wgpu::CommandEncoderDescriptor {
-                                    label: Some("Copy Encoder"),
-                                });
-
-                            // Copy the output buffer to the texture
-                            copy_encoder.copy_buffer_to_texture(
-                                wgpu::ImageCopyBuffer {
-                                    buffer: &output_buffer,
-                                    layout: wgpu::ImageDataLayout {
-                                        offset: 0,
-                                        bytes_per_row: Some(bytes_per_row), // 4 bytes per pixel
-                                        rows_per_image: Some(size.height),
-                                    },
-                                },
-                                wgpu::ImageCopyTexture {
-                                    texture: &texture,
-                                    mip_level: 0,
-                                    origin: wgpu::Origin3d::ZERO,
-                                    aspect: wgpu::TextureAspect::All,
-                                },
-                                wgpu::Extent3d {
-                                    width: size.width,
-                                    height: size.height,
-                                    depth_or_array_layers: 1,
-                                },
+                            scene_renderer.update_frame(
+                                &device,
+                                &queue,
+                                &compute_pipeline,
+                                &compute_bind_group,
+                                &texture,
                             );
-
-                            queue.submit(Some(copy_encoder.finish()));
 
                             // #############################################################################################
 
@@ -791,33 +577,6 @@ async fn run(event_loop: EventLoop<()>, window: Window) {
             }
         })
         .unwrap();
-}
-
-fn update_render_queue(
-    queue: &wgpu::Queue,
-    texture: &wgpu::Texture,
-    size: &winit::dpi::PhysicalSize<u32>,
-    pixel_colors: &[u8],
-) {
-    queue.write_texture(
-        wgpu::ImageCopyTexture {
-            texture,
-            mip_level: 0,
-            origin: wgpu::Origin3d::ZERO,
-            aspect: wgpu::TextureAspect::All,
-        },
-        pixel_colors,
-        wgpu::ImageDataLayout {
-            offset: 0,
-            bytes_per_row: Some(size.width * 4),
-            rows_per_image: Some(size.height),
-        },
-        wgpu::Extent3d {
-            width: size.width,
-            height: size.height,
-            depth_or_array_layers: 1,
-        },
-    );
 }
 
 fn create_texture(device: &wgpu::Device, size: winit::dpi::PhysicalSize<u32>) -> wgpu::Texture {
@@ -1001,285 +760,6 @@ fn generate_instance() -> Instance {
 }
 
 // ######################### compute ########################################
-
-fn generate_camera_rays(size: &winit::dpi::PhysicalSize<u32>) -> Vec<Ray> {
-    let mut ray_vec: Vec<Ray> = vec![];
-
-    for y in 0..size.height {
-        let y = (y as f32 / size.height as f32) * 2.0 - 1.0;
-        for x in 0..size.width {
-            let x = (x as f32 / size.width as f32) * 2.0 - 1.0;
-
-            let ray = Ray {
-                direction: vec3_pad(x, y, -1.0),
-            };
-
-            ray_vec.push(ray);
-        }
-    }
-
-    ray_vec
-}
-
-fn generate_new_camera_rays(camera: &Camera) -> Vec<Ray> {
-    let mut ray_vec: Vec<Ray> = vec![];
-
-    let rays = camera.ray_directions.clone();
-
-    for ray in rays {
-        let direction = ray.direction;
-        let new_ray = Ray {
-            direction: vec3_pad(direction.x, direction.y, direction.z),
-        };
-
-        ray_vec.push(new_ray)
-    }
-
-    ray_vec
-}
-
-fn create_buffers(
-    device: &wgpu::Device,
-    size: &winit::dpi::PhysicalSize<u32>,
-    camera_rays: &[Ray],
-    material_array: &[SceneMaterial],
-    sphere_array: &[SceneSphere],
-    params: &[Params],
-) -> (
-    u64,
-    u64,
-    Buffer,
-    Buffer,
-    Buffer,
-    Buffer,
-    Buffer,
-    Buffer,
-    Buffer,
-    Buffer,
-) {
-    let input_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-        label: Some("Input Buffer"),
-        contents: bytemuck::cast_slice(camera_rays),
-        usage: wgpu::BufferUsages::STORAGE | wgpu::BufferUsages::COPY_DST,
-    });
-
-    // 4 bytes of u8 per pixel, RGBA
-    let output_buffer_size =
-        (size.width * size.height * 4 * std::mem::size_of::<u8>() as u32) as wgpu::BufferAddress;
-
-    let output_buffer = device.create_buffer(&wgpu::BufferDescriptor {
-        label: Some("Output Buffer"),
-        size: output_buffer_size,
-        usage: wgpu::BufferUsages::STORAGE | wgpu::BufferUsages::COPY_SRC,
-        mapped_at_creation: false,
-    });
-
-    let params_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-        label: Some("Params Buffer"),
-        contents: bytemuck::cast_slice(params),
-        usage: wgpu::BufferUsages::STORAGE | wgpu::BufferUsages::COPY_DST,
-    });
-
-    // Create staging buffer for reading back data
-    let staging_buffer = device.create_buffer(&wgpu::BufferDescriptor {
-        label: Some("Staging Buffer"),
-        size: output_buffer_size,
-        usage: wgpu::BufferUsages::MAP_READ | wgpu::BufferUsages::COPY_DST,
-        mapped_at_creation: false,
-    });
-
-    // Create uniform buffer
-    let camera = RayCamera {
-        origin: vec3_pad(-7.0, -7.0, 25.),
-        direction: vec3_pad(0.0, 0.0, -1.0),
-    };
-
-    let camera_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-        label: Some("Camera Buffer"),
-        contents: bytemuck::cast_slice(&[camera]),
-        usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-    });
-
-    let material_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-        label: Some("Material Buffer"),
-        contents: bytemuck::cast_slice(material_array),
-        usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-    });
-
-    let sphere_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-        label: Some("Sphere Buffer"),
-        contents: bytemuck::cast_slice(sphere_array),
-        usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-    });
-
-    // 3 bytes of f32 per pixel
-    let accumulation_buffer_size =
-        (size.width * size.height * std::mem::size_of::<[f32; 4]>() as u32) as wgpu::BufferAddress;
-
-    let accumulation_buffer = device.create_buffer(&wgpu::BufferDescriptor {
-        label: Some("Accumulation Buffer"),
-        size: accumulation_buffer_size,
-        usage: wgpu::BufferUsages::STORAGE | wgpu::BufferUsages::COPY_DST,
-        mapped_at_creation: false,
-    });
-
-    (
-        output_buffer_size,
-        accumulation_buffer_size,
-        input_buffer,
-        output_buffer,
-        params_buffer,
-        staging_buffer,
-        camera_buffer,
-        material_buffer,
-        sphere_buffer,
-        accumulation_buffer,
-    )
-}
-
-fn create_compute_bindgroup(
-    device: &wgpu::Device,
-    input_buffer: &Buffer,
-    output_buffer: &Buffer,
-    params_buffer: &Buffer,
-    camera_buffer: &Buffer,
-    material_buffer: &Buffer,
-    sphere_buffer: &Buffer,
-    accumulation_buffer: &Buffer,
-) -> (wgpu::BindGroupLayout, BindGroup) {
-    let params_bind = 0;
-    let ray_directions_bind = 1;
-    let pixel_colors_bind = 2;
-    let camera_bind = 3;
-    let material_bind = 4;
-    let sphere_bind = 5;
-    let accumulation_bind = 6;
-
-    let bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-        entries: &[
-            wgpu::BindGroupLayoutEntry {
-                binding: params_bind,
-                visibility: wgpu::ShaderStages::COMPUTE,
-                ty: wgpu::BindingType::Buffer {
-                    ty: wgpu::BufferBindingType::Storage { read_only: true },
-                    has_dynamic_offset: false,
-                    min_binding_size: None,
-                },
-                count: None,
-            },
-            wgpu::BindGroupLayoutEntry {
-                binding: ray_directions_bind,
-                visibility: wgpu::ShaderStages::COMPUTE,
-                ty: wgpu::BindingType::Buffer {
-                    ty: wgpu::BufferBindingType::Storage { read_only: true },
-                    has_dynamic_offset: false,
-                    min_binding_size: None,
-                },
-                count: None,
-            },
-            wgpu::BindGroupLayoutEntry {
-                binding: pixel_colors_bind,
-                visibility: wgpu::ShaderStages::COMPUTE,
-                ty: wgpu::BindingType::Buffer {
-                    ty: wgpu::BufferBindingType::Storage { read_only: false },
-                    has_dynamic_offset: false,
-                    min_binding_size: None,
-                },
-                count: None,
-            },
-            wgpu::BindGroupLayoutEntry {
-                binding: camera_bind,
-                visibility: wgpu::ShaderStages::COMPUTE,
-                ty: wgpu::BindingType::Buffer {
-                    ty: wgpu::BufferBindingType::Uniform,
-                    has_dynamic_offset: false,
-                    min_binding_size: None,
-                },
-                count: None,
-            },
-            wgpu::BindGroupLayoutEntry {
-                binding: material_bind,
-                visibility: wgpu::ShaderStages::COMPUTE,
-                ty: wgpu::BindingType::Buffer {
-                    ty: wgpu::BufferBindingType::Uniform,
-                    has_dynamic_offset: false,
-                    min_binding_size: None,
-                },
-                count: None,
-            },
-            wgpu::BindGroupLayoutEntry {
-                binding: sphere_bind,
-                visibility: wgpu::ShaderStages::COMPUTE,
-                ty: wgpu::BindingType::Buffer {
-                    ty: wgpu::BufferBindingType::Uniform,
-                    has_dynamic_offset: false,
-                    min_binding_size: None,
-                },
-                count: None,
-            },
-            wgpu::BindGroupLayoutEntry {
-                binding: accumulation_bind,
-                visibility: wgpu::ShaderStages::COMPUTE,
-                ty: wgpu::BindingType::Buffer {
-                    ty: wgpu::BufferBindingType::Storage { read_only: false },
-                    has_dynamic_offset: false,
-                    min_binding_size: None,
-                },
-                count: None,
-            },
-        ],
-        label: None,
-    });
-
-    let compute_bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-        layout: &bind_group_layout,
-        entries: &[
-            wgpu::BindGroupEntry {
-                binding: params_bind,
-                resource: params_buffer.as_entire_binding(),
-            },
-            wgpu::BindGroupEntry {
-                binding: ray_directions_bind,
-                resource: input_buffer.as_entire_binding(),
-            },
-            wgpu::BindGroupEntry {
-                binding: pixel_colors_bind,
-                resource: output_buffer.as_entire_binding(),
-            },
-            wgpu::BindGroupEntry {
-                binding: camera_bind,
-                resource: camera_buffer.as_entire_binding(),
-            },
-            wgpu::BindGroupEntry {
-                binding: material_bind,
-                resource: material_buffer.as_entire_binding(),
-            },
-            wgpu::BindGroupEntry {
-                binding: sphere_bind,
-                resource: sphere_buffer.as_entire_binding(),
-            },
-            wgpu::BindGroupEntry {
-                binding: accumulation_bind,
-                resource: accumulation_buffer.as_entire_binding(),
-            },
-        ],
-        label: None,
-    });
-
-    (bind_group_layout, compute_bind_group)
-}
-
-fn calculate_bytes_per_row(size: &winit::dpi::PhysicalSize<u32>) -> u32 {
-    // the gpu buffer has to be 256 * n bytes per row
-
-    let bytes_per_pixel = 4; // RGBA8Unorm = 4 bytes per pixel
-
-    let value = size.width * bytes_per_pixel;
-    let alignment = wgpu::COPY_BYTES_PER_ROW_ALIGNMENT;
-
-    // bytes per row
-    (value + alignment - 1) & !(alignment - 1)
-}
 
 fn create_ui(platform: &mut Platform, screne_renderer: &mut Renderer) -> FullOutput {
     platform.begin_frame();
